% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_repo_data}
\alias{get_repo_data}
\title{Get repository data for the specified organization and repo groups}
\usage{
get_repo_data(org, groups)
}
\arguments{
\item{org}{Organization name.}

\item{groups}{Repo groups, see \code{\link{repo_group}}.}
}
\description{
Get repository data for the specified organization and repo groups
}

