% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{myghrepos}
\alias{myghrepos}
\title{Create a Web Page About Your GitHub Repositories}
\usage{
myghrepos(org, target = tempfile(), groups = NULL, data = NULL)
}
\arguments{
\item{org}{GitHub organization or user name.}

\item{target}{Target directory for the generated file(s).}

\item{groups}{Repo groups, see \code{\link{repo_group}}.}

\item{data}{The repo data, if you already have it, as
returned by \code{\link{get_repo_data}}.}
}
\description{
This is a static website generator, specialized on creating a single
page web site about GitHub repositories, for an account or organization.
}

