% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ myghrepos }
\alias{ myghrepos }
\alias{ myghrepos }
\title{ Create a Nice Web Page About Your GitHub Repositories }
\description{
This is a static website generator, specialized on creating a single page web
  site about GitHub repositories, for an account or organization.
}
