% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_group}
\alias{repo_group}
\title{Define a group of repositories}
\usage{
repo_group(name, description = NULL, repos)
}
\arguments{
\item{name}{Group name.}

\item{description}{Group description.}

\item{repos}{Character vector, repositories to include in the group.}
}
\value{
Repo group object.
}
\description{
Define a group of repositories
}

