% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{safe_whisker}
\alias{safe_whisker}
\title{Workaround for whisker mishandling corner cases}
\usage{
safe_whisker(template, ...)
}
\arguments{
\item{template}{Template.}

\item{...}{Extra arguments to \code{whisker.render}.}
}
\description{
Workaround for whisker mishandling corner cases
}
\keyword{internal}

